v1.6.13 - Add event log rate limiter to catch misbehaving plugins

+++ src/main/java/me/danjono/inventoryrollback/data/YAML.java
@@ -201,9 +201,13 @@ public class YAML {
             timeSaved.add(saveTimeStamp);
         }
 
+
         for (int i = 0; i < deleteAmount; i++) {
+            if (timeSaved.isEmpty()) break;
+
             Long deleteTimestamp = Collections.min(timeSaved);
             timeSaved.remove(deleteTimestamp);
+
             try {
                 Files.deleteIfExists(new File (playerBackupFolder, deleteTimestamp + ".yml").toPath());
             } catch (IOException ex) {
